and if func exists for custom logo

+++ header-front.php
@@ -41,9 +41,16 @@
                 <div class="logo-container">
                     <a class="home-logo-link" href="<?php echo get_home_url(); ?>">
                         <div alt="Logo" class="logo-content logo-front-page img-fluid">
+
+                        <?php if( function_exists('custom_logo')) { ?>
+
                             <img class="logo-img" src="<?php $custom_logo_id = get_theme_mod( 'custom_logo' );
                                 $image = wp_get_attachment_image_src( $custom_logo_id , 'full' );
                                     echo $image[0]; ?>"></img>
+
+                        <?php }else{} ?>
+
+                            
                             <h1 class="mb-2 h3 logo-heading"><?php echo get_bloginfo('name'); ?></h1>
                         </div>
                     </a>
